requirements.txt
faiss-cpu
gradio
openai
pypdf
numpy
